refactor: extract all Fn:: handlers into modular files under src/lib/functions/
Phase 1 complete:
- Created src/lib/functions/ with types, helpers, and individual handler files
- fn-array-ops.ts: Flatten, FlattenDeep, Uniq, Compact, Concat, Sort, SortedUniq, SortBy, Without
- fn-object-ops.ts: Omit, OmitEmpty, Merge, DeepMerge, ObjectKeys, ObjectValues, SortObject
- fn-string-ops.ts: Stringify, StringSplit, UpperCamelCase, LowerCamelCase, JoinNow, SubNow
- fn-misc.ts: GetEnv, Eval, Filenames, Outputs, Sequence, IfEval, RefNow, ApplyTags
- fn-map.ts: Map handler
- fn-length.ts: Length handler
- fn-include.ts: Include handler with fnInclude, handleIncludeBody
- registry.ts: dispatch map Record<string, FnHandler>
- helpers.ts: findAndReplace, interpolate, JSONifyString, getBoolEnvOpt
- src/index.ts reduced from 889 to ~100 lines, clean dispatcher
- All 197 tests pass
